// src/app/menubar/models/deliveryaddress.model.ts
export class deliveryaddress {
    addressID:   number =0;
    orderID:     number=0;
    addressName: string='test';
    mobileNo:    string='00';
    zipCode:     string='55';
    locality:    string='tr';
    address:     string='tr';
    state:       string='ts';
    country:     string='ts';
    landmark:    string='miya';
    alternateNo: string='543';
}